selinux: use arrays for avtab hashtable nodes
The current avtab hashtable employs a separate chaining collision
resolution strategy where each bucket/chain holds an ordered linked list
of pointers to kmem_cache allocated avtab_node elements.

On Fedora 38 (x86_64) using the default policy, avtab_node_cachep
uses 573 slabs each containing 170 objects totaling 2,337,840 bytes.
A call to kmem_cache_zalloc() is required for every single rule, which
in the default policy is currently 96,730 and continually rising.

When both sets of avtab_node (regular and cond.) are turned into arrays
with the hash table slot indexing into it rather than a pointer, then
this results in only 2 additional kvcalloc() calls and the complete
removal of the kmem_cache itself.

Due to how conditional rules are written in the binary policy, the
code responsible for loading does not know how many conditional rules
there are before creating the avtab structure. Instead, it uses the
number of elements in the non-conditional avtab as a hint and allocates
the hash table based on it. Therefore, pre-allocating an array of nodes
based on this hint over-allocates at best and under-allocates at worst.

This patch includes functions which help manage the size of the nodes
array when there are more or less conditional rules than regular rules.
Because we are using pointers, pointer math is used to rebase them on
the new node array. These functions include avtab_grow_nodes() and
avtab_shrink_nodes() in avtab.c and cond_shrink_av() and
cond_rebase_av_list() in conditional.c. Future improvements to the
binary policy to provide the correct hint to the loader code will make
these functions obsolete and removable. But as this would be a breaking
change to the policy format, it is not a part of this patch.

The caching characteristics of iterating a single array are better due
to locality of reference. Running "perf stat -r 100 -d load_policy"
has shown a runtime reduction of at least 10% on a Fedora 38 x86_64 VM
with this single patch. Future patches focused on improving the hash
table's collision resolution strategy and array layout (struct-of-arrays
vs. array-of-structs) may elicit even more caching and therefore runtime
performance improvements.

Signed-off-by: Jacob Satterfield <jsatterfield.linux@gmail.com>